Moved functions back to 'kit.ts'
