+++ src/installs/visual-studio.ts
@@ -44,28 +44,7 @@ interface VSInstallationCache {
   queryTime: number;
 }
 
-
 let cachedVSInstallations: VSInstallationCache|null = null;
-
-
-/**
- * Construct the display name (this will be paired with an
- * arch later to construct the Kit.name property).
- *
- * @param inst The VSInstallation to use
- */
-export function vsDisplayName(inst: VSInstallation): string {
-  if (inst.displayName) {
-    if (inst.channelId) {
-      const index = inst.channelId.lastIndexOf('.');
-      if (index > 0) {
-        return `${inst.displayName} ${inst.channelId.substr(index + 1)}`;
-      }
-    }
-    return inst.displayName;
-  }
-  return inst.instanceId;
-}
 
 /**
  * Get a list of all Visual Studio installations available from vswhere.exe.
@@ -107,12 +86,4 @@ export async function vsInstallations(): Promise<VSInstallation[]> {
     queryTime: now
   };
   return installs;
-}
-
-export function vsVersionName(inst: VSInstallation): string {
-  if (!inst.catalog) {
-    return inst.instanceId;
-  }
-  const end = inst.catalog.productDisplayVersion.indexOf('[');
-  return end < 0 ? inst.catalog.productDisplayVersion : inst.catalog.productDisplayVersion.substring(0, end - 1);
 }
